experimenting with an options system provided by the backend itself
Each backend shall provide class attributes ''description'' and ''options'' - ''descriptions'' being just a string, ''options'' is a list of option dictionaries.

The values of these dicts are:

 ''option'':: name of that option
 ''type'':: string or int
 ''default'':: default value, should be of type ''type'' (optional)
 ''enum'':: a tuple/list of possible values (optional)
 ''range'':: a dict with ''min'' and ''max'' elements (optional)
 ''help'':: a (short) text describing this option (optional)
 ''level'':: the level of this option - ''normal'' or ''advance'' (optional)

Anything else?

Accessing these attributes is like that:

{{{
>>> from coherence.base import Plugins
>>> p = Plugins()
>>> p['FSStore'].options
[{'default': 'my media', 'type': 'string', 'key': 'name'}, ...]
}}}

The Plugins class is a dict, so p.items() returns the backend identifiers as keys and their classes as values.

git-svn-id: https://coherence.beebits.net/svn/trunk/Coherence@1444 8402a193-bf1e-0410-a7e5-8246bfde4f7d